+++ src/init.cpp
@@ -72,7 +72,7 @@
 #include <stdio.h>
 #include <memory>
 
-#include "crypto/bls.h"
+#include "bls/bls.h"
 
 #ifndef WIN32
 #include <signal.h>
